persistence: Use mongodb as datastore
The repositories in the erp and the warehouse are now using mongodb
as datastore to save the queue data.

+semver: minor

+++ src/LogisticsSuite.Backend/Hubs/IMonitorHub.cs
@@ -10,13 +10,13 @@ namespace LogisticsSuite.Backend.Hubs
 
 		Task OnDelayChangedAsync(DelayDto delay);
 
-		Task OnOrderQueueChangedAsync(int count);
+		Task OnOrderQueueChangedAsync(long count);
 
 		Task OnOrderReleasedMessageReceivedAsync(OrderReleasedMessage message);
 
 		Task OnParcelDispatchedMessageReceivedAsync(ParcelDispatchedMessage message);
 
-		Task OnParcelQueueChangedAsync(int count);
+		Task OnParcelQueueChangedAsync(long count);
 
 		Task OnReplenishedMessageReceivedAsync(ReplenishedMessage message);
